More unittests updates (WIP).
