+++ tests/mocks/api.github.com/repos/mozilla-services/shavar/956c36c9a470c152abf34c7ae247292e987ede2c/CHANGES.txt
