+++ tests/mocks/api.github.com/repos/mozilla-services/shavar/git/refs/tags
